requirements.txt
RPi.GPIO
requests
